+++ src/style/Detail/HeaderInDetail.style.jsx
@@ -1,0 +1,38 @@
+import styled from 'styled-components';
+import { theme } from '../theme';
+
+const S = {};
+
+S.DetailContainer = styled.div`
+  width: 100%;
+  height: 70vh;
+  margin: 50px auto;
+  padding: 20px;
+  display: flex;
+  flex-direction: column;
+  background-color: ${({ theme }) => theme.colors.nav_background_color};
+  border-radius: 30px;
+  box-sizing: border-box;
+`;
+
+S.HeaderInDetail = styled.div`
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  /* background-color: blue; */
+  /* color: white; */
+  padding: 10px 20px;
+  color: #424242;
+  border-radius: 30px;
+  margin-bottom: 0px;
+  box-sizing: border-box;
+`;
+
+S.TitleContainer = styled.div`
+  box-sizing: border-box;
+  font-weight: bold;
+  font-size: 20px;
+`;
+
+export default S;
